requirement.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
dj-rest-auth==7.0.1
Django==5.2.6
django-allauth==65.11.1
django-cors-headers==4.7.0
djangorestframework==3.16.1
idna==3.10
pillow==11.3.0
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
